backend/app/requirements.txt
fastapi
uvicorn
langchain
openai
chromadb
python-dotenv
transformers
google-generativeai==0.3.2
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3